Fix additional merge conflicts
