+++ src/competitor.py
@@ -1,11 +1,9 @@
 #!/usr/bin/env python3
 
 # helper
+import math
 import random
 
-import numpy as np
-
-# own files
 import utils as ut
 
 
@@ -15,49 +13,48 @@ class Competitor:
         self.maxquality = ut.MAX_QUALITY
         self.quality = ut.shuffle_quality()
 
-    def get_initial_price(self, random_start):
-        if random_start:
-            return max(
-                int(ut.PRODUCTION_PRICE + np.random.normal() * 3 + 3),
-                ut.PRODUCTION_PRICE + 1,
-            )
+    def reset(self):
+        return (
+            # 1 will be added to the price because profit is unpossible else
+            random.randint(ut.PRODUCTION_PRICE + 1, ut.MAX_PRICE),
+            ut.shuffle_quality(),
+        )
+
+    def give_competitors_price(self, state, self_idx):
+        assert False, 'You must use a subclass of Competitor!'
+
+
+class CompetitorLinearRatio1(Competitor):
+    def give_competitors_price(self, state, self_idx):
+        # this stratgy calculates the value per money for each competing vendor and tries to adept to it
+        ratios = []
+        max_competing_ratio = 0
+        for i in range(math.floor(len(state) / 2)):
+            ratio = state[2 * i + 1] / state[2 * i]  # value for money for vendor i
+            ratios.append(ratio)
+            if ratio > max_competing_ratio and i != self_idx:
+                max_competing_ratio = ratio
+
+        ratio = max_competing_ratio / ratios[self_idx]
+        randomnumber = random.random()
+        intended = 0
+        if randomnumber < 0.15:
+            intended = state[2 * self_idx] + random.randint(-1, 1)
         else:
-            return ut.PRODUCTION_PRICE
+            intended = math.floor(1 / max_competing_ratio * state[2 * self_idx + 1])
+        return min(min(max(ut.PRODUCTION_PRICE + 1, intended), ut.MAX_PRICE), 23)
 
-    def reset(self, random_start):
-        self.quality = ut.shuffle_quality()
-        return self.get_initial_price(random_start), self.quality
 
-    def give_competitors_price(self, state):
-        return self.comp_strategy1(state)
+class CompetitorRandom(Competitor):
+    def give_competitors_price(self, state, _):
+        return random.randint(ut.PRODUCTION_PRICE + 1, ut.MAX_PRICE)
 
-    def comp_strategy1(self, state):
-        # this stratgy is based on a price quality ratio
-        agent_price = state[0]
-        comp_price = state[2]
-        agent_quality = state[1]
-        comp_quality = state[3]
 
-        ratio = (agent_quality / agent_price) / (comp_quality / comp_price)
-        if random.random() < 0.1:
-            return comp_price + random.randint(-1, 1)
-        elif comp_price < ut.PRODUCTION_PRICE or (
-            ratio < 0.95 and comp_price < ut.MAX_PRICE - 5
-        ):
-            return agent_price + 1
-        elif (
-            agent_price > ut.PRODUCTION_PRICE + 1
-            and ratio > 1.05
-            or comp_price > ut.MAX_PRICE - 5
-        ):
-            return agent_price - 1
-        elif agent_price > ut.PRODUCTION_PRICE + 1 and comp_quality > agent_quality:
-            return agent_price + 1
-        else:
-            return comp_price
-
-    def comp_strategy2(self, state):
-        # this competitor is based on quality and agents actions
+class CompetitorJust2Players(Competitor):
+    def give_competitors_price(self, state, _):
+        # This competitor is based on quality and agents actions.
+        # While he can act in every linear economy, you should not expect good performance in a multicompetitor setting.
+        # assert len(state) == 4, "You can't use this competitor in this market!"
         agent_price = state[0]
         comp_price = state[2]
         agent_quality = state[1]
