+++ src/human_player.py
@@ -1,27 +1,22 @@
 #!/usr/bin/env python3
 
 # rl
-# from gym.spaces.space import Space
-
 # own files
 import utils as ut
-from sim_market import SimMarket
+from sim_market import ClassicScenario
 
-env = SimMarket()
+env = ClassicScenario()
 our_profit = 0
 is_done = False
 state = env.reset()
 
 
-print('The production price is', ut.PRODUCTION_PRICE)
+print('The production price: ', ut.PRODUCTION_PRICE, ' maxprice: ', ut.MAX_PRICE)
 while not is_done:
-    agent_price = state[0]
-    comp_price = state[2]
-    agent_quality = state[1]
-    comp_quality = state[3]
+    agent_quality = state[0]
+    comp_price = state[1]
+    comp_quality = state[2]
     print(
-        'agent_price:',
-        agent_price,
         'agent_quality',
         agent_quality,
         'comp_price:',
@@ -30,8 +25,7 @@ while not is_done:
         comp_quality,
     )
     action = input('What do you want to do? ')
-    if action == '':
-        action = agent_price
+    assert action != ''
     state, reward, is_done, _ = env.step(int(action))
     print('Your profit this round is', reward)
     our_profit += reward
