--- tests/test_customers.py
@@ -1,13 +1,0 @@
-# from ..sim_market import Customer
-
-
-# ident_state = [15, 20, 15, 20]
-
-# We need a test so the CI doesn't fail...
-def test_because_we_have_to():
-    assert 1 == 1
-
-
-# # Check that Customers only choose actions between 0 and 2
-# def test_customer_action_range():
-#     assert 0 <= Customer.buy_object(ident_state) <= 2
